Move the spool-side inventory tests beside the code they test
tests/test_proc/test_proc_inventory.py kept only what tests
Patch.enrich; its spool classes -- attach/remove, lazy reads and
equality, enrichment on extraction, inventory-backed select/unselect,
conform, and expand_by -- move to tests/test_core/test_spool_inventory.py.
The classes named after review rounds (TestReviewFindings and friends)
are renamed for the behavior they pin, and their tests sit with the
class that owns that behavior: the union tests under TestSpoolUnion, the
selector edge shapes under TestSelectorEdgeShapes, the channel-selection
contracts under TestChannelSelectContracts.
